safari-534.53-branch/LayoutTests/platform/chromium-win/http/tests/loading/bad-scheme-subframe-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
frame "f1" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
frame "f1" - didFailProvisionalLoadWithError
main frame - didFinishLoadForFrame
This is a test of load callbacks. It is only useful inside the regression test tool.
